// src/app/models/enums/toast-notification-type.enum.ts
export enum ToastNotificationType {
  IMPORTING_FILE_INSTEAD_OF_FOLDER_ERROR = 'IMPORTING_FILE_INSTEAD_OF_FOLDER_ERROR',
  CASE_IMPORTED_SUCCESSFULLY = 'CASE_IMPORTED_SUCCESSFULLY',
  CASE_EXPORTED_TO_FOLDER_SUCCESSFULLY = 'CASE_EXPORTED_TO_FOLDER_SUCCESSFULLY',
  CASE_EXPORTED_TO_DRIVE_SUCCESSFULLY = 'CASE_EXPORTED_TO_DRIVE_SUCCESSFULLY',
  CASE_EXPORTED_TO_DISK_SUCCESSFULLY = 'CASE_EXPORTED_TO_DISK_SUCCESSFULLY',
  FILE_EXPORTED_TO_CASE_FOLDER_LOCATION_SUCCESSFULLY = 'FILE_EXPORTED_TO_CASE_FOLDER_LOCATION_SUCCESSFULLY',
  DESIGN_PARAMETERS_UPDATED_SUCCESSFULLY = 'DESIGN_PARAMETERS_UPDATED_SUCCESSFULLY',
  PROFILE_DETAILS_UPDATED_SUCCESSFULLY = 'PROFILE_DETAILS_UPDATED_SUCCESSFULLY',
  CASE_NAME_UDATED_SUCCESSFULLY = 'CASE_NAME_UDATED_SUCCESSFULLY',
  CASE_DELETED_SUCCESSFULLY = 'CASE_DELETED_SUCCESSFULLY',
  AVATAR_UPDATED_SUCCESSFULLY = 'AVATAR_UPDATED_SUCCESSFULLY',
  PASSWORD_UPDATED_SUCCESSFULLY = 'PASSWORD_UPDATED_SUCCESSFULLY',
  EMAIL_UPDATED_SUCCESSFULLY = 'EMAIL_UPDATED_SUCCESSFULLY',
  SELECTED_DENTAL_CASE_NOT_FOUND_IN_IMPORT_FOLDER_ERROR = 'SELECTED_DENTAL_CASE_NOT_FOUND_IN_IMPORT_FOLDER_ERROR',
  PERMISSION_TO_FOLDER_NOT_GRANTED_ERROR = 'PERMISSION_TO_FOLDER_NOT_GRANTED_ERROR',
  CANNOT_START_NEW_CASE_AFTER_RESTART_ERROR = 'CANNOT_START_NEW_CASE_AFTER_RESTART_ERROR',
  SIDEBAR_PATIENT_ID_COPIED_SUCCESSFULLY = 'SIDEBAR_PATIENT_ID_COPIED_SUCCESSFULLY',
  OTHER = 'OTHER',
}